fix: precios decimales, unidad de medida, modal de busqueda, estabilidad
PROBLEMAS ARREGLADOS:

1. PRECIOS DECIMALES (el problema principal)
   - El sistema usaba int64 (enteros) para todos los precios
   - 2 TQ / 5 productos = 0.4 TQ se redondeaba a 0 TQ
   - Cambiado a float64 / NUMERIC(12,2) en todo el sistema
   - Migracion 046_decimal_prices.sql cambia BIGINT a NUMERIC(12,2)
   - Archivos cambiados:
     * internal/pricing/calculator.go (todos los campos de energia y precio)
     * internal/external/store.go (PriceTrueque, BasePrice, ExtraCosts, FinalPrice)
     * internal/external/dex.go (LocalPriceTrueque)
     * internal/api/external.go (CompositeComponent, StoreItem, listComponents)
     * internal/api/handlers.go (CreateProductRequest, CalcInternalRequest)
     * internal/api/system.go (var price float64 en todos los lugares)
     * internal/federation/protocol.go (PricePerUnit)
     * internal/federation/server.go (BroadcastProductProposal)
   - Frontend: Math.round() eliminado, ahora usa toFixed(2) para mostrar decimales

2. UNIDAD DE MEDIDA DEL PRODUCTO COMPUESTO
   - Antes: el producto compuesto siempre era "unidad" con cantidad 1
   - Ahora: el usuario selecciona la unidad (ml, L, gr, kg, unidad, manojo, etc.)
   - Ahora: el usuario especifica la cantidad por unidad (200ml, 500gr, etc.)
   - Esto permite saber exactamente cuanto contiene cada producto
   - Ejemplo: Jugo de naranja, unidad=ml, cantidad=200
   - El backend recibe unit y quantity_per_unit en el request

3. MODAL DE BUSQUEDA DE COMPONENTES
   - Antes: mostraba "Escribe para buscar componentes" cuando no habia texto
   - Ahora: carga la lista completa de componentes al abrir el modal
   - Anadido estado de loading y error para diagnosticar problemas
   - Si la sesion expira, muestra mensaje claro en vez de lista vacia
   - Si el backend falla, muestra el error en vez de silenciosamente vaciar

4. ESTABILIDAD DEL BACKEND
   - Timeout del middleware aumentado de 30s a 60s
   - middleware.Recoverer ya estaba presente (atra panics)
   - Manejo de errores mas robusto en listComponents

5. CALCULO DE PRECIOS EN FRONTEND
   - compositeTotalPrice ahora usa decimales sin redondear
   - Display de componentes usa toFixed(2) en vez de Math.round
   - Display de precio total usa toFixed(2)

+++ internal/api/handlers.go
@@ -353,11 +353,11 @@ type CreateProductRequest struct {
 	Origin             string   `json:"origin"`
 	Unit               string   `json:"unit"`
 	QuantityPerBatch   int      `json:"quantity_per_batch"`
-	EnergyDirect       int64    `json:"energy_direct"`
-	EnergyHuman        int64    `json:"energy_human"`
-	EnergyInputs       int64    `json:"energy_inputs"`
-	EnergyAmortization int64    `json:"energy_amortization"`
-	PricePerUnit       int64    `json:"price_per_unit"`
+	EnergyDirect       float64  `json:"energy_direct"`
+	EnergyHuman        float64  `json:"energy_human"`
+	EnergyInputs       float64  `json:"energy_inputs"`
+	EnergyAmortization float64  `json:"energy_amortization"`
+	PricePerUnit       float64  `json:"price_per_unit"`
 	ExternalPriceUSD   *float64 `json:"external_price_usd"`
 	ExternalTaxRate    float64  `json:"external_tax_rate"`
 	Description        string   `json:"description"`
@@ -410,11 +410,11 @@ func (h *Handler) getTariff(w http.ResponseWriter, r *http.Request) {
 
 type CalcInternalRequest struct {
 	Quantity           int     `json:"quantity"`
-	EnergyDirect       int64   `json:"energy_direct"`
+	EnergyDirect       float64 `json:"energy_direct"`
 	HoursHuman         float64 `json:"hours_human"`
 	LaborType          string  `json:"labor_type"`
-	EnergyInputs       int64   `json:"energy_inputs"`
-	EnergyAmortization int64   `json:"energy_amortization"`
+	EnergyInputs       float64 `json:"energy_inputs"`
+	EnergyAmortization float64 `json:"energy_amortization"`
 }
 
 func (h *Handler) calcInternal(w http.ResponseWriter, r *http.Request) {
